+++ performance_test/src/experiment_execution/runner.hpp
@@ -87,11 +87,11 @@ public:
       }
 
       if (ignore_first_seconds_of_experiment(experiment_start)) {
-        auto results = std::make_shared<AnalysisResult>();
-        results->m_experiment_start = loop_start - experiment_start;
-        results->m_time_between_two_measurements =
+        AnalysisResult results;
+        results.m_experiment_start = loop_start - experiment_start;
+        results.m_time_between_two_measurements =
           time_between_two_measurements;
-        results->m_cpu_info = cpu_usage_tracker.get_cpu_usage();
+        results.m_cpu_info = cpu_usage_tracker.get_cpu_usage();
 
         for (auto & pub : m_pub_stats) {
           pub.populate_stats(results);