+++ performance_test/src/outputs/stdout_output.hpp
@@ -31,7 +31,7 @@ public:
   virtual ~StdoutOutput() = default;
 
   void open() override;
-  void update(std::shared_ptr<const AnalysisResult> result) override;
+  void update(const AnalysisResult & result) override;
   void close() override;
 
 private: