+++ performance_test/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 - Update the README to better explain how to use this tool with Apex.OS
+- In the `Runner`, allocate the `AnalysisResult`s on the stack instead of using `shared_ptr`
 
 ## 1.4.2 (2023/03/15)